+++ run.py
@@ -1,5 +1,6 @@
 from flask import render_template
 
+from blueprints.admin import admin_bp
 from blueprints.auth import auth_bp
 from blueprints.tests import test_bp
 from blueprints.user import user_bp
@@ -7,6 +8,7 @@ from main import create_app
 
 app = create_app()
 app.register_blueprint(user_bp, url_prefix="/user")
+app.register_blueprint(admin_bp, url_prefix="/admin")
 app.register_blueprint(auth_bp)
 app.register_blueprint(test_bp)
